pyproject.toml
[tool.poetry]
name = "sc-mapmerge"
version = "2.1.0"
description = "Utility for merging stalcraft map into single image"
authors = ["onejeuu <bloodtrail@beber1k.ru>"]
license = "MIT"
packages = [{ include = "scmapmerge" }]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
pillow = "^10.2.0"
rich = "^13.7.0"
sc-file = "^2.0.0"
click = "^8.1.7"
inquirerpy = "^0.3.4"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.4.0"

[tool.poetry.scripts]
build = "build:build"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
